requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
echo==0.8.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
mypy==1.9.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pandas-stubs==2.2.1.240316
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
QtPy==2.4.1
requests==2.31.0
requests-mock==1.11.0
ruff==0.3.4
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.36.3
tomli==2.0.1
types-psycopg2==2.9.21.20240311
types-pytz==2024.1.0.20240203
types-requests==2.31.0.20240311
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
